New UI Implementation (#316)
* added new health bar icon

* updated pause menu

changed the pause box + added the graffiti + changed the image for audio/sensitivity handles. i rearranged the order of the buttons as well to match similarly to the mock up.

* final pause box changes

changed the font + added a few missing details of the pause box to match the mock up. made it a new prefab called "NewPauseMenu."

* updated Icon + pause menu

- new radial gague icon

---------

Co-authored-by: TheStress <lieu.nelson386@gmail.com>